ufs: don't set fid version to mtime
Using mtime as the Version incorrectly shows a file being a different file when it's not.

Instead, set it to 0. Yes, a file that is changed to a different file *might* look the same
if they ever get the same inumber. This false positive is not near as harmful as the
false negative we had before.

It's a shame the Go runtime doesn't let us see CreateTime.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>