Tstat/Twstat: correctly create packets.
This is an old weird thing in 9p, but what is carried int stat and wstat
is not known to the protocol. So ALL of us, over the years, who decoded stat
in the protocol code were wrong. In the protocol, stat/wstat stat structs
are just a []byte, with -- ah fun -- a 16-bit length preceding them, not
32-bit as in most things.

so we resort to the hack of a type called DataCnt16, which is used by the stub
generator to create correct marshal/unmarshal code. The call code is still []byte
because there's no reason to expose that nonsense to the programmer.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>