xfilesystem: correctly return an error when a walk element fails.
If a client walks a/b/c the server must return an array of QIDS
for all three elements. If the walk fails at any point the
server must return an error as well as an array of as many
QIDS as it was able to walk.

Example: if only a/b exists, and the client tries to walk
a/b/c, the server must return the QIDs for a and b as well
as an error.

Somehow I forgot this. This fixes a problem we'd seen in harvey
that was not a harvey problem. #294.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>