filesystem: return an array of QIDS one longer than the last successful walk
For as long as I've been writing 9p servers (20 years now, oh joy)
I thought a walk that ended in an
error should only return the number of QIDs successfully walked. Oops.
It needs to return one more, the last one being invalid;
unless the number is zero, in which case it should return error and no
QIDS.

This fixes an error message printing problem seen in harvey.

So far, the score is harvey 3, Go servers 0, when it comes to correctness in 9p :-)
This server has had two errors so far, found in simple ls commands.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>